+++ packages/datawell/src/react/__stories__/index.story.tsx
@@ -1,10 +1,7 @@
 import { Meta, Story } from '@storybook/react/types-6-0'
-import glamorDefault, * as glamorExports from 'glamor'
 import React from 'react'
 
 import DataWell from '../index'
-
-const glamor = glamorDefault || glamorExports
 
 export default {
   title: 'Components/DataWell',
@@ -33,23 +30,28 @@ export const Row: Story = () => (
 
 export const FixedWidthRow: Story = () => (
   <div
-    {...glamor.css({
-      display: 'flex',
-      '& > div': {
-        width: '25%'
-      }
-    })}
+    style={{
+      display: 'flex'
+    }}
   >
-    <DataWell label="Dog count" subLabel="All the doggies">
-      234,345
-    </DataWell>
-    <DataWell label="Cat count">123</DataWell>
-    <DataWell label="Rafters on the River">1,345/23,235</DataWell>
-    <DataWell
-      label="Flotsam"
-      subLabel="The icky stuff that is not cats or dogs, floating in the river"
-    >
-      About 3
-    </DataWell>
+    <div style={{ width: '25%' }}>
+      <DataWell label="Dog count" subLabel="All the doggies">
+        234,345
+      </DataWell>
+    </div>
+    <div style={{ width: '25%' }}>
+      <DataWell label="Cat count">123</DataWell>
+    </div>
+    <div style={{ width: '25%' }}>
+      <DataWell label="Rafters on the River">1,345/23,235</DataWell>
+    </div>
+    <div style={{ width: '25%' }}>
+      <DataWell
+        label="Flotsam"
+        subLabel="The icky stuff that is not cats or dogs, floating in the river"
+      >
+        About 3
+      </DataWell>
+    </div>
   </div>
 )