+++ packages/icon/src/index.ts
@@ -1,4 +1,4 @@
-export * from './vars'
-export { default } from './react'
+export { default } from './react/index'
 export * from './react/icons'
 export * as icons from './react/icons'
+export * from './vars/index'