+++ packages/textinput/src/index.ts
@@ -1,2 +1,3 @@
-export { default, TextInputProps } from './react/index'
+export { default } from './react/index'
+export * from './react/index'
 export * from './vars/index'