+++ packages/checkbox/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-checkbox@10.0.1...@pluralsight/ps-design-system-checkbox@10.0.2) (2021-08-12)
+
+
+### Bug Fixes
+
+* **checkbox:** rm glamor references ([9172041](https://github.com/pluralsight/design-system/commit/9172041cff29ca282228edeaf4fedaf1e8c91a36))
+
+
+
+
+
 ## [10.0.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-checkbox@10.0.0...@pluralsight/ps-design-system-checkbox@10.0.1) (2021-07-27)